+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.2 branch
 ### 2.2.8
+* fixed bug for thumbnails for urls with query string;
 * updated to CakePHP 3.2.
 
 ### 2.2.7